backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
transformers==4.36.2
torch==2.1.2
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.3
sentencepiece==0.2.0
